Update zone_eversong_woods.sql
Quarelestra <Cooking Trainer>

This is a TBC npc, it's debatable if this one should be changed.
It was decided to include the bloodelf starting zones to the vanilla experience, but I think starting out here should not give an advantage so that's why I removed the ability for this npc to train expert and artisan cooking. Together with removing the spiced bread recipe, which is too convenient for a vanilla cooking recipe.