fix pvp Items requiring Honor Ranks
You were only able to equip pvp items requiring your exact rank 
because the player's pvp title is checked and previously earned pvp titles are overwritten. This resulted in the player not being able to equip items that require a lower pvp rank.

Now higher ranks are also checked.

The required rank text on the items remains red
no solution for that yet.