Delete optional/dbc.7z
In this version I changed the spell.dbc file to set the level requirement to 20 for ghost wolf and aspect of the cheetah.

I don't want to do that anymore because it may mess up bots using the spells. The level 20 requirement is still there at the trainer.

This does mean that you could add the spell manually as a GM to use it at level 16.