update comments AGAIN
The progression phases are very confusing.

From tests:
with RequirePreAQQuests set to 0 the gate is open during the pre aq phase. But you should't get to this phase normally because the Nefarius kill will place you in the AQ phase.
You skip the war effort normally with RequirePreAQQuests = 0

But I guess it's fine like this, because now you can set yourself manually to the pre-AQ phase if you  want to experience the war effort.

with RequirePreAQQuests set to 1 the AQ gate remains closed during the AQ phase. you need to do the quest. if you manually set yourself to the NAXX phase the gate opens up as well..