Update individualProgression.conf.dist
removed RDF setting.
it did nothing. there was no code for it.

there already is a setting for the RDF in worldserver conf
people can disable the RDF there.

including it in the SimpleConfigOverride complicates things
because if you want to enable the RDF you change several other things as well. That you likely don't want to change.