Remove RDF from SimpleConfigOverride config option (#561)
removed RDF from the SimpleConfigOverride
including it in the SimpleConfigOverride complicates things
because if you wanted to enable the RDF you were forced to change several other settings as well.

it's all very confusing
I'm trying to make it simple.

With disableRDF in the Individual Progression config file you can now enable/disable the RDF.
that's it. no need to edit the worldserver.conf file for this any more.

it's also very clear now when it's disabled. can't even select dungeons anymore.